% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{adjust_factor}
\alias{adjust_factor}
\title{Applies an importance weight to an ipfr factor}
\usage{
adjust_factor(factor, importance)
}
\arguments{
\item{factor}{A correction factor that is calculated using target/current.}

\item{importance}{A \code{real} between 0 and 1 signifying the importance of
the factor. An importance of 1 does not modify the factor. An importance of
0.5 would shrink the factor closer to 1.0 by 50 percent.}
}
\value{
The adjusted factor.
}
\description{
At lower values of importance, the factor is moved closer to 1.
}
\keyword{internal}
