% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{balance_secondary_targets}
\alias{balance_secondary_targets}
\title{Balances secondary targets to primary}
\usage{
balance_secondary_targets(primary_targets, primary_seed, secondary_targets,
  secondary_seed)
}
\arguments{
\item{primary_targets}{Input from \link{ipu} after scaling.}

\item{secondary_targets}{Input from \link{ipu} after scaling.}
}
\value{
\code{named list} of the secondary targets
}
\description{
The average weight per record needed to satisfy targets is computed for both
primary and secondary targets. Often, these can be very different, which leads
to poor performance. The algorithm must use extremely large or small weights
to match the competing goals. The secondary targets are scaled so that they
are consistent with the primary targets on this measurement.
}
\details{
If multiple geographies are present in the secondary_target table, then
balancing is done for each geography separately.
}
