% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{check_geo_fields}
\alias{check_geo_fields}
\title{Check geo fields}
\usage{
check_geo_fields(seed, target, target_name)
}
\arguments{
\item{seed}{seed table to check}

\item{target}{data.frame of a single target table}

\item{target_name}{the name of the target (e.g. size)}
}
\value{
The seed and target table (which may be modified)
}
\description{
Helper function for \code{\link{check_tables}}. Makes sure that geographies
in a seed and target table line up properly.
}
\keyword{internal}
