% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{check_missing_categories}
\alias{check_missing_categories}
\title{Check for missing categories in seed}
\usage{
check_missing_categories(seed, target, target_name, geo_colname)
}
\arguments{
\item{seed}{seed table to check}

\item{target}{data.frame of a single target table}

\item{target_name}{the name of the target (e.g. size)}

\item{geo_colname}{the name of the geo column in both the \code{seed} and
\code{target} (e.g. geo_taz)}
}
\value{
Nothing. Throws an error if one is found.
}
\description{
Helper function for \code{check_tables}.
}
\keyword{internal}
