% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{check_seed}
\alias{check_seed}
\title{Check for complete seed table}
\usage{
check_seed(seed, targets)
}
\arguments{
\item{seed}{\code{data.frame} Seed table}

\item{targets}{\code{list} Target tables}
}
\description{
Given seed and targets, checks to make sure that at least one
  observation of each marginal category exists in the seed table.  Otherwise,
  ipf/ipu would produce wrong answers without throwing errors.
}
