% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{check_tables}
\alias{check_tables}
\title{Check seed and target tables for completeness}
\usage{
check_tables(hh_seed, hh_targets, per_seed, per_targets)
}
\arguments{
\item{hh_seed}{Same as \code{seed} in \link[ipfr]{ipf}. Each row is a
household. Must contain an \code{hhid} column.}

\item{hh_targets}{Same as \code{targets} in \link[ipfr]{ipf}.}

\item{per_seed}{Seed table of person attributes. Each row is a person.
Must also contain an \code{hhid} column that links each person to their
respective household in \code{hh_seed}.}

\item{per_targets}{Similar to \code{targets} from \link[ipfr]{ipf}, but for
person-level targets.}
}
\description{
Given seed and targets, checks to make sure that at least one
  observation of each marginal category exists in the seed table.  Otherwise,
  ipf/ipu would produce wrong answers without throwing errors.
}
