% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{compare_results}
\alias{compare_results}
\title{Compare results to targets}
\usage{
compare_results(seed, targets)
}
\arguments{
\item{seed}{\code{data.frame} Seed table with a weight column in the same
format required by \code{ipu()}.}

\item{targets}{\code{named list} of \code{data.frames} in the same format
required by \code{ipu()}.}
}
\description{
Compare results to targets
}
