% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh}
\alias{hh}
\title{Household travel survey records}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	1026 obs. of  11 variables:
 $ houseid   : chr  "36587836" "51591788" "53703464" "61244177" ...
 $ county    : chr  "161" "161" "161" "161" ...
 $ hh_weight : num  74.3 75.2 57.6 54.4 59.3 ...
 $ vehicles  : num  3 1 1 3 2 1 2 3 1 3 ...
 $ adults    : num  3 2 2 2 2 2 2 3 1 2 ...
 $ workers   : num  2 0 0 0 0 1 0 2 1 0 ...
 $ drivers   : num  3 2 1 2 1 2 2 3 1 2 ...
 $ income    : chr  "14" "10" "04" "13" ...
 $ size      : num  3 2 2 2 2 2 2 3 2 2 ...
 $ income_grp: Factor w/ 2 levels "Under $50k","More than $50k": 2 1 1 2 2 2 1 NA 2 1 ...
 $ sufficient: logi  FALSE FALSE FALSE FALSE FALSE TRUE ...
}}
\usage{
hh
}
\description{
\code{hh} is a reduced and cleaned version of nhts records from Roanoke
county, Virginia.
}
\keyword{datasets}

