% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh_joint}
\alias{hh_joint}
\title{Joint Household Distribution}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 64 rows and 4 columns.}
\usage{
hh_joint
}
\description{
\code{hh_joint} is the joint distribution of households by size, workers, 
and vehicles.
}
\keyword{datasets}
