% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{ipf}
\alias{ipf}
\title{Reweight a Seed Table to Marginal Controls}
\usage{
ipf(seed, weight_var = NULL, variables, marginals, relative_gap = 0.01,
  max_iterations = 50, verbose = FALSE)
}
\arguments{
\item{seed}{a \code{data.frame} including necessary colums for matching to 
marginals and an optional weight field.}

\item{weight_var}{existing weights column in seed table. 
If \code{NULL}, defaults to 1.}

\item{variables}{a character vector of the table variable names that will
match against marginal tables.}

\item{marginals}{a list of \code{data.frame} objects, each a marginal table corresponding 
to \code{variables}. The tables should have columns corresponding to 
possible values for the appropriate \code{var*} variable.}

\item{verbose}{Print the maximum expansion factor with each iteration? 
Default \code{FALSE}.}
}
\value{
a vector of weights for each row in \code{seed}
}
\description{
Reweight a Seed Table to Marginal Controls
}

