% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfr.R
\name{ipf_multi}
\alias{ipf_multi}
\title{Perform the ipf procedure for multiple marginal sets and returns a 
\code{data.frame}.}
\usage{
ipf_multi(seed, weight_var = "weight", margTbl, id_field,
  relative_gap = 0.01, absolute_gap = 1, max_iterations = 50,
  min_weight = 1e-04, verbose = FALSE)
}
\arguments{
\item{seed}{a \code{data.frame} including necessary colums for matching to 
marginals and an optional weight field.}

\item{weight_var}{existing weights column in seed table. 
If \code{NULL}, defaults to 1.}

\item{margTbl}{A \code{data.frame} with a row for each set of marginals to
use in an ipf procedure.  Every column other than the \code{id_field} will
be treated as a marginal column.}

\item{id_field}{The name of the identifying field not to be used as a 
marginal.  IPF will be performed for each ID.}

\item{relative_gap}{defines convergence.  If if no cells are factored by more 
than this amount, the process is said to have converged.  For example, the
default value of .01 imlies that convergence is reached when no cell in 
seed table is changed by more than 10 percent.}

\item{absolute_gap}{The relative gap checks the percent difference.  The 
absolute gap checks the absolute difference.  If the absolute difference is 
below the absolute_gap, the percent difference is ignored.}

\item{max_iterations}{maximimum number of iterations to perform, even if 
convergence is not reached.}

\item{min_weight}{Minimum weight to allow in any cell to prevent zero weights.
Set to .0001 by default.  Should be arbitrarily small compared to your 
seed table weights.}

\item{verbose}{Print details on the maximum expansion factor with each 
iteration? Default \code{FALSE}.}
}
\value{
a \code{data.frame} with a row for each \code{id_field}. The columns 
   of which contain the joint-distribution weights after fitting to marginals.
}
\description{
Perform the ipf procedure for multiple marginal sets and returns a 
\code{data.frame}.
}

