% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\docType{package}
\name{ipfr}
\alias{ipfr}
\alias{ipfr-package}
\title{ipfr: A package to perform iterative proportional fitting}
\description{
There are currently three functions
}
\details{
\code{\link{ipf}} (deprecated)

\code{\link{ipu}}

\code{\link{ipu_nr}}

Both \code{ipu} and \code{ipu_nr} implement list balancing. \code{ipu_nr}
implements a newton-raphson approach to balance primary and secondary
targets. \code{ipu} balances targets directly, which can mean faster
convergence.
}
