% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{ipu}
\alias{ipu}
\title{Iterative Proportional Updating}
\usage{
ipu(hh_seed, hh_targets, per_seed, per_targets, damp_factor = 0.75,
  relative_gap = 0.01, max_iterations = 100, absolute_diff = 10,
  min_weight = 0.0001, verbose = FALSE)
}
\arguments{
\item{hh_seed}{Same as \code{seed} in \link[ipfr]{ipf}. Each row is a
household. Must contain an \code{hhid} column.}

\item{hh_targets}{Same as \code{targets} in \link[ipfr]{ipf}.}

\item{per_seed}{Seed table of person attributes. Each row is a person.
Must also contain an \code{hhid} column that links each person to their
respective household in \code{hh_seed}.}

\item{per_targets}{Similar to \code{targets} from \link[ipfr]{ipf}, but for
person-level targets.}

\item{damp_factor}{Should be a number between 0 and 1. Reduces the adjustment
factor calculated in each iteration before updating weights. Lower values
require more iterations but are less likely to overfit a single marginal
distribution at the expense of the others. Defaults to 0.75.}

\item{relative_gap}{After each iteration, the weights are compared to the
previous weights and an RMSE metric is calculated. If the RMSE is less than
the \code{relative_gap} threshold, then the process terminates.}

\item{max_iterations}{maximimum number of iterations to perform, even if 
\code{relative_gap} is not reached.}

\item{absolute_diff}{Upon completion, the \code{ipu()} function will report
the worst-performing marginal category and geography based on the percent
difference from the target. \code{absolute_diff} is a threshold below which
percent differences don't matter.

For example, if if a target value was 2, and the expanded weights equaled
1, that's a 100% difference, but is not important because the absolute value
is only 1.

Defaults to 10.}
}
\value{
the \code{hh_seed} with a revised weight column.
}
\description{
A special case of iterative proportional fitting when two,
disparate sets of marginals need to be satisfied that do not agree on a
single total. A common example is balance population data using household-
and person-level marginal controls. This could be for survey expansion or
synthetic population creation. The original Arizona IPU method is improved
by dampening the adjustment factor.
}
\references{
\url{http://www.scag.ca.gov/Documents/PopulationSynthesizerPaper_TRB.pdf}
}
