% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{process_seed_table}
\alias{process_seed_table}
\title{Helper function to process a seed table}
\usage{
process_seed_table(df, primary_id, marginal_columns)
}
\arguments{
\item{df}{the \code{data.frame} as processed by \code{ipu()} before this
function is called.}

\item{primary_id}{the name of the primary ID column.}

\item{marginal_columns}{The vector of column names in the seed table that
have matching targets.}
}
\description{
Helper for \code{ipu()}. Strips columns from seed table except for the
primary id and marginal column (as reflected in the targets tables). Also
identifies factor columns with one level and processes them before
\code{mlr::createDummyFeatures()} is called.
}
\keyword{internal}
