% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand.R
\name{reweight_survey}
\alias{reweight_survey}
\title{Reweight a Survey to Marginal Controls}
\usage{
reweight_survey(survey, weight_var = NULL, variables, marginals,
  relative_gap = 0.01, max_iterations = 50, verbose = FALSE)
}
\arguments{
\item{survey}{a \code{data.frame} with survey responses, including
  necessary colums for matching to marginals.}

\item{weight_var}{existing weights variable. If \code{NULL}, defaults to 1.}

\item{variables}{a character vector of the variables \code{survey} that will
match against marginal tables.}

\item{marginals}{a list of \code{data.frame} objects, each a marginal table corresponding
to \code{variables}. The tables should have rows for each \code{geo_var} and
columns corresponding to possible values for the appropriate \code{var*}
variable.}

\item{verbose}{Print the maximum expansion factor with each iteration?
  Default \code{FALSE}.}
}
\value{
a vector of weights for each row in \code{survey}
}
\description{
Reweight a Survey to Marginal Controls
}

