% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{scale_targets}
\alias{scale_targets}
\title{Scale targets to ensure consistency}
\usage{
scale_targets(targets, verbose = FALSE)
}
\arguments{
\item{targets}{\code{named list} of \code{data.frames} in the same format
required by \link{ipu}.}

\item{verbose}{\code{logical} Show a warning for each target scaled?
Defaults to \code{FALSE}.}
}
\value{
A \code{named list} with the scaled targets
}
\description{
Often, different marginals may disagree on the total number of units. In the
context of household survey expansion, for example, one marginal might say
there are 100k households while another says there are 101k. This function
solves the problem by scaling all target tables to match the first target
table provided.
}
\keyword{internal}
