% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_arizona}
\alias{setup_arizona}
\title{Create the ASU example}
\usage{
setup_arizona()
}
\value{
A list of four variables:
  hh_seed, hh_targets, per_seed, and per_targets. These can be used directly
  by \code{\link{ipu}}.
}
\description{
Sets up the Arizona example IPU problem and is used in multiple places
throughout the package (vignettes/tests).
}
\examples{
setup_arizona()
}
