% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{Creates a synthetic population based on ipu results}
\usage{
synthesize(weight_tbl, geo_field = NULL, primary_id = "id")
}
\arguments{
\item{weight_tbl}{the \code{data.frame} of the same name output by
\code{\link{ipu}}.}

\item{geo_field}{if provided, the \code{weight_tbl} will be grouped
by that field before the random sampling is done.}

\item{primary_id}{The field used to join the primary and secondary seed
tables. Only necessary if \code{secondary_seed} is provided.}
}
\value{
A \code{data.frame} with one record for each synthetized member of
  the population (e.g. household). A \code{new_id} column is created, but
  the previous \code{primary_id} column is maintained to facilitate joining
  back to other data sources (e.g. a person table).
}
\description{
A simple function that takes the \code{weight_tbl} output from
\code{\link{ipu}} and randomly samples based on the weight.
}
