% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipfr.R
\docType{package}
\name{ipfr}
\alias{ipfr}
\alias{ipfr-package}
\title{ipfr: A package to perform iterative proportional fitting}
\description{
The main function is \code{\link{ipu}}. For a 2D/matrix problem, the 
\code{\link{ipu_matrix}} function is easier to use. The resulting
\code{weight_tbl} from \code{ipu()} can be fed into \code{\link{synthesize}}
to generate a synthetic population
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dkyleward/ipfr}
  \item Report bugs at \url{https://github.com/dkyleward/ipfr/issues}
}

}
\author{
\strong{Maintainer}: Kyle Ward \email{kyleward084@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Greg Macfarlane \email{gregmacfarlane@byu.edu} [contributor]
}

}
