% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\docType{package}
\name{ipfr}
\alias{ipfr}
\alias{ipfr-package}
\title{ipfr: A package to perform iterative proportional fitting}
\description{
There are currently two functions
}
\details{
\code{\link{ipu}}

\code{\link{ipf}}

The \code{ipf} function is deprecated. \code{ipu} has a more consistent
input format, can handle primary and secondary marginals, and can use varying
geographical constraints for each set of targets.

Use \code{ipu}
}
