% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R, R/ipu_nr.R
\docType{package}
\name{ipfr}
\alias{ipfr}
\alias{ipfr-package}
\alias{ipfr}
\alias{ipfr-package}
\title{ipfr: A package to perform iterative proportional fitting}
\description{
There are currently two functions

There are currently two functions
}
\details{
\code{\link{ipu}}

\code{\link{ipf}}

The \code{ipf} function is deprecated. \code{ipu} has a more consistent
input format, can handle primary and secondary marginals, and can use varying
geographical constraints for each set of targets.

Use \code{ipu}

\code{\link{ipu}}

\code{\link{ipu_nr}}

Both functions implement list balancing. \code{ipu_nr} implements a
newton-raphson approach to balance primary and secondary targets. \code{ipu}
balances targets directly, which can mean faster convergence.
}
