% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\docType{package}
\name{ipfr}
\alias{ipfr}
\alias{ipfr-package}
\title{ipfr: A package to perform iterative proportional fitting}
\description{
The main function is \code{\link{ipu}}. For a 2D/matrix problem, the 
\code{\link{ipu_matrix}} function is easier to use.
}
