% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu.R
\name{ipu}
\alias{ipu}
\title{Iterative Proportional Updating}
\usage{
ipu(hh_seed, hh_targets, per_seed, per_targets, relative_gap = 0.01,
  absolute_gap = 1, max_iterations = 50, min_weight = 0.0001,
  verbose = FALSE)
}
\arguments{
\item{hh_seed}{Same as \code{seed} in \link[ipfr]{ipf}. Each row is a
household. Must contain an \code{hhid} column.}

\item{hh_targets}{Same as \code{targets} in \link[ipfr]{ipf}.}

\item{per_seed}{Seed table of person attributes. Each row is a person.
Must also contain an \code{hhid} column that links each person to their
respective household in \code{hh_seed}.}

\item{per_targets}{Similar to \code{targets} from \link[ipfr]{ipf}, but for
person-level targets.}

\item{relative_gap}{target for convergence.  Maximum percent change to allow
any seed weight to move by while considering the process converged.  By 
default, if no weights change by more than 1%, the process has converged.
The process is said to be converged if either \code{relative_gap} or 
\code{absolute_gap} parameters have been met.}

\item{absolute_gap}{target for convergence.  Maximum absolute change to allow
any seed weight to move by while considering the process converged.  By 
default, if no weights change by more than 10, the process has converged.
The process is said to be converged if either \code{relative_gap} or 
\code{absolute_gap} parameters have been met.}

\item{max_iterations}{maximimum number of iterations to perform, even if 
convergence is not reached.}

\item{min_weight}{Minimum weight to allow in any cell to prevent zero weights.
Set to .0001 by default.  Should be arbitrarily small compared to your 
seed table weights.}

\item{verbose}{Print details on the maximum expansion factor with each 
iteration? Default \code{FALSE}.}
}
\value{
the \code{hh_seed} with a revised weight column.
}
\description{
A special case of iterative proportional fitting when two,
disparate sets of marginals need to be satisfied that do not agree on a
single total. A common example is balance population data using household-
and person-level marginal controls. This could be for survey expansion or
synthetic population creation.
}
\references{
\url{http://www.scag.ca.gov/Documents/PopulationSynthesizerPaper_TRB.pdf}
}
